+++ commands/CDeveloper.js
@@ -71,6 +71,22 @@ class CDeveloper {
         });
     }
 
+    _deletePayment(token, options, prefs) {
+
+        var id = prefs.user.id;
+        if(options.id) {
+            id = options.id;
+        }
+
+        return new Promise(function(resolve, reject) {
+            NodeSDK.put('/api/rainbow/subscription/v1.0/developers/' + id + '/accounts/cancel', token, null).then(function(json) {
+                resolve(json);
+            }).catch(function(err) {
+                reject(err);
+            });
+        });
+    }
+
     getPayments(options) {
         var that = this;
 
@@ -196,6 +212,55 @@ class CDeveloper {
                 Message.error(err, options);
                 Exit.error();
             });
+        }
+        else {
+            Message.notLoggedIn(options);
+            Exit.error();
+        }
+    }
+
+    deletePayment(options) {
+        var that = this;
+
+        var doDelete = function(options) {
+            Message.action("Delete payment", options.id, options);
+
+            let spin = Message.spin(options);
+            NodeSDK.start(that._prefs.email, that._prefs.password, that._prefs.host).then(function() {
+                Message.log("execute action...");
+                return that._deletePayment(that._prefs.token, options, that._prefs);
+            }).then(function(json) {
+                Message.unspin(spin);
+                Message.log("action done...", json);
+                Message.lineFeed();
+                Message.success(options);
+                Message.log("finished!");
+            }).catch(function(err) {
+                Message.unspin(spin);
+                Message.error(err, options);
+                Exit.error();
+            });
+        }
+        
+        Message.welcome(options);
+                
+        if(this._prefs.token && this._prefs.user) {
+            Message.loggedin(this._prefs, options);
+
+            if(options.noconfirmation) {
+                doDelete(options);
+            }
+            else {
+                Message.confirm('Are-you sure ? It will remote it completely').then(function(confirm) {
+                    if(confirm) {
+                        doDelete(options);
+                    }
+                    else {
+                        Message.canceled(options);
+                        Exit.error();
+                    }
+                });
+            }
         }
         else {
             Message.notLoggedIn(options);